LPS-65804 Move
