+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-test/src/test/java/com/liferay/dynamic/data/mapping/type/key/value/internal/KeyValueDDMFormFieldTypeSettingsTest.java
@@ -12,13 +12,12 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.type.key.value;
+package com.liferay.dynamic.data.mapping.type.key.value.internal;
 
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.model.DDMFormRule;
 import com.liferay.dynamic.data.mapping.type.BaseDDMFormFieldTypeSettingsTest;
-import com.liferay.dynamic.data.mapping.type.key.value.internal.KeyValueDDMFormFieldTypeSettings;
 import com.liferay.dynamic.data.mapping.util.DDMFormFactory;
 import com.liferay.portal.kernel.util.PortalClassLoaderUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
